ADR-0016 sweep P2: rename committed-access -> sealed-access (no logic change)
Pure rename, kept in its own commit so the diff is reviewable: the module
`src/lib/evidence/committed-access.ts` becomes `sealed-access.ts` and its
`isCommittedRecord` predicate becomes `isSealedRecord`, per ADR-0016 §A's
state-label rename (`committed` -> `sealed`). The delegation to
`isSealedDbValue` in the vocabulary boundary module is unchanged, as are all
call sites' semantics — only the import path and the exported name move.

The "commitment" noun stays put everywhere (ADR-0016 §A's razor): this module
gates CONTENT disclosure, which is the state label, not the cryptographic
commitment.
